refactor code and add docblocks

+++ database/migrations/2022_02_04_001622_create_feature_plan_table.php
@@ -20,7 +20,7 @@ return new class() extends Migration
       $table->foreignIdFor(config('larasubs.models.plan'))
         ->cascadeOnDelete();
 
-      $table->decimal('units')->nullable();
+      $table->integer('units')->nullable();
 
       $table->timestamps();
